requirements.txt
#base requirements
Flask>=0.12.1
flask-mongoengine>=0.9.2
Flask-Script>=2.0.5
itsdangerous>=0.24
mock>=2.0.0
redis>=2.10.5
Werkzeug>=0.12.1
uwsgi>=2.0.15
flask-cors>=3.0.2
bcrypt>=3.1.3


#test dependencies
pytest>=3.0.7
pytest-cov>=2.4.0
requests-mock>=1.3.0
